pyproject.toml
[project]
name = "nonebot-plugin-hermes"
version = "0.1.0"
description = "NoneBot plugin for Hermes Agent — multi-platform AI chatbot via Hermes API Server"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

dependencies = [
    "nonebot2>=2.3.0",
    "nonebot-plugin-alconna>=0.50.2",
    "httpx>=0.27.0",
    "cachetools>=6.2.0",
]

[project.urls]
"Homepage" = "https://github.com/gsskk/nonebot-plugin-hermes"

[tool.nonebot]
plugins = ["nonebot_plugin_hermes"]

[tool.ruff]
line-length = 120
exclude = ["references"]

[tool.ruff.lint]
ignore = ["E402"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.setuptools]
packages = ["nonebot_plugin_hermes"]
